requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
markdown2>=2.5.4
gunicorn==21.2.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.14.0